services/web/requirements.txt
APScheduler==3.10.4
asgiref
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gunicorn==21.2.0
Jinja2==3.1.2
nats-py==2.10.0
Pillow==10.0.0
psycopg2-binary==2.9.9
pydantic-settings==2.0.2
python-dotenv==1.0.0
python-telegram-bot==20.2
SQLAlchemy==2.0.20
Werkzeug==2.3.7
WTForms==3.0.1